requirements.txt
### Data Science & ML
lightgbm
numpy
pandas
scikit-learn
shap

### Deep Learning & NLP
accelerate
datasets
sentence-transformers
torch
transformers

### Data Visualisation
matplotlib
seaborn
wordcloud
Pillow

### APIs & Hubs
huggingface_hub
openai

### Demo App
gradio

### Utilities
nltk
spacy
tqdm
python-dotenv